+++ src/index.ts
@@ -1,28 +1,19 @@
 import { JsonTsMapperService } from './lib/json-ts-mapper.service';
-import { JsonObject, JsonConverter, JsonProperty, Optional, NotNull, CustomConverter } from './lib/decorators';
-import { Context, AbstractJsonConverter } from './lib/json-converter';
-import { DateTimeConverter } from './lib/date-converters'
+import { JsonProperty, Optional, NotNull, Converter } from './lib/decorators';
+import { AbstractJsonConverter } from './lib/json-converter';
 import { Any } from './lib/any';
 
 const JsonTsMapper = new JsonTsMapperService();
 
-const DateConverter = DateTimeConverter
-
 export {
   JsonTsMapper,
 
-  JsonObject,
-  JsonConverter,
   JsonProperty,
   Optional,
   NotNull,
-  CustomConverter,
+  Converter,
 
   Any,
 
-  Context,
-  AbstractJsonConverter,
-
-  DateTimeConverter,
-  DateConverter
+  AbstractJsonConverter
 };