+++ src/lib/json-converter.ts
@@ -3,8 +3,6 @@ export type ConverterInstance = AbstractJsonConverter<any, any>;
 
 export abstract class AbstractJsonConverter<JsonType extends string | number | boolean | { [key: string]: any }, TsType>{
 
-  abstract serialize (obj: TsType, context?: Context): JsonType;
-  abstract deserialize (obj: JsonType, context?: Context): TsType;
+  abstract serialize (obj: TsType): JsonType;
+  abstract deserialize (obj: JsonType): TsType;
 }
-
-export type Context = { [name: string]: any };