+++ changelog.md
@@ -1,3 +1,15 @@
+## 3.0.0
+ * remove @JsonObject
+ * remove @JsonConverter
+ * remove context in custom converters
+ * rework metadata
+ * rework build
+ * fix types
+ * fix decorators : use all with ()
+
+## 2.2.1
+ * change visibility of 'desiarializeObject' and 'deserializeArray' to public
+
 ## 2.2.0
  * add possibility to use object or class instance as converter